Rename Trapframe to Stackframe due to conflict
This commit renames Trapframe in lab1 to Stackframe due to a naming conflict in lab3.

The author originally misunderstood Trapframe in monitor.h to be the "Stackframe" structure that would be used for stackwalk and thus gave it an incorrect definition when it should have been left alone.

This change rectifies that by reverting Trapframe to be a just declaration and defining a new Stackframe struct.