This completes lab4
This completes lab4 by implementing IPC.
Also fixes a bug in env_run where before switching away
from curenv, its status is indiscriminately set to RUNNABLE,
even in cases where curenv should be "blocked"(such as
in case of waiting to receive IPC).

This leads to a race condition in ipc_recv where the
syscall number for IPC is returned in eax register because
the sender env has not set the receiver env's eax to 0
yet.

The fix is to change env_run so that curenv's status is
set to RUNNABLE iff its status is RUNNING.